added missing method to MetricAccessor protocol
